slight modification of Tasker schema / add useful relationships
- The dependence relationship between two tasks contained a
redundant functional dependency : the two tasks must be in
the same project. This is fixed by having a shared project_id
column, participating in both foreign key constraints. I do not
know if this is the correct approach or if I just did something
hacky and the the whole model is bad

- A Contributor is now related to a project (and not just identified
by a member). This is  more practical as we will always be interacting
with a contributor in the context of a specified project.