[debug] add timeout wrapper around where the task hangs
On last logs it seems that the task hangs sometimes just before
awaiting a fetch_message call. With some tests I discovered
that this does not necessarily mean the task is hanging but
can also be caused by an uncatched exception.
I don't know why an uncatched exception is not reported by the
task runner and why it causes it to stop running it again, and
I should investigate this. For the time being I added a timeout
arount the fetch_message call to see if it really hangs, and also
added a clause to catch any unexpected exception and report on it.